requirements.txt
# Core dependencies
setuptools>=68.0.0
wheel>=0.40.0
build>=1.0.0
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
azure-storage-blob>=12.0.0
python-dotenv>=1.0.0
python-multipart>=0.0.6

# Core AI/ML Dependencies (Realistic 3D Blinds)
opencv-python-headless>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0
scipy>=1.11.0
scikit-image>=0.21.0
azure-cognitiveservices-vision-computervision>=0.9.0
google-generativeai>=0.3.0
requests>=2.31.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.24.0

# Rate limiting
slowapi>=0.1.9 